backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy>=2.2.0
opencv-python-headless==4.12.0.88
pillow==11.0.0
psycopg2-binary==2.9.10
PyMySQL==1.1.2
requests==2.32.3
SQLAlchemy==2.0.43
typing_extensions>=4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==23.0.0
torch>=2.5.1
torchvision
transformers
timm
reportlab==4.2.5
